refactor: extract explicit nested runtime
Move nested shape, preflight, migration, pruning, and projection into a
cycle-safe subsystem. Keep mixed decorator/cardinality orchestration in
the facade.

All 33 definitions are source- and AST-identical to origin/main.
Preserve canonical excluded-field filtering.
Retain the shared missing-value sentinel identity.

Validation: 858 tests passed on Pydantic 2.13.4 (95.86% coverage).
Validation: 858 tests passed on Pydantic 2.12.3 (95.77% coverage).
Validation: 540 focused tests and 38 independent boundaries passed.
Validation: Ruff, ty, strict mypy, Vulture, and strict docs passed.
Validation: immutable v0.3.0 contract passed.

Closes #114
Refs #94
